## Summary
Convert all physical inventory quantities from integer types to BigDecimal and
PostgreSQL NUMERIC(19,4), following the Release 1 domain precision contract.

## Changes

- Convert item balances, sale quantities, receipt quantities, adjustment
  quantities, stock-movement balances, and audit quantities to BigDecimal.
- Propagate BigDecimal through API DTOs, services, repositories, mappers,
  dashboard aggregation, low-stock projections, and CSV/XLSX handling.
- Add centralized inventory quantity validation.
- Use compareTo for numeric decisions and avoid quantity conversion through double.
- Calculate sale subtotals directly from BigDecimal quantities.
- Add Flyway V7 to convert all persisted physical quantities to NUMERIC(19,4).
- Add compatible positive and nonnegative database constraints.

## Validation

- Incoming quantities must be positive.
- Input scale may not exceed four, including trailing decimal zeros.
- Inputs are never silently rounded.
- Non-fractional items require mathematically whole quantities.
- Resulting stock may not be negative.

## Migration

Existing integer values are preserved exactly as scale-four decimals.
Adjustment change_quantity remains signed and therefore intentionally has no
nonnegative constraint.

## Tests

mvn verify passes across all eight modules with 14 tests and no failures.

## Out of scope

No UOM conversion, roll-to-meter conversion, stock-transfer functionality,
cash-session behavior, or unrelated monetary-field changes.

+++ bloom-app-api/src/main/java/com/bloom/app/api/dto/response/dashboard/LowStockDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class LowStockDto {
     private Long id;
     private String name;
     private String sku;
-    private Integer stock;
-    private Integer minStock;
+    private BigDecimal stock;
+    private BigDecimal minStock;
     private String category;
 }
