+++ bloom-app-api/src/main/java/com/bloom/app/api/dto/response/stockadjustment/CsvParseResponse.java
@@ -6,13 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class CsvParseResponse {
     private String itemSku;
-    private Integer changeQuantity;
+    private BigDecimal changeQuantity;
     private StockAdjustmentActionType actionType;
     private String reason;
 }
